test: split attachments.test.js
